Make ScoverageKeys a top level object
